Revert Wine update to the version we have in Stretch.
TI Nspire CX CAS license activation does not work
(with our patch) with the version in Buster.
Downgrade Wine to the version in Stretch.

This reverts commits cff8e43b2ffffe67c747355757a1dbbae3a18706,
75c8092c5663c528d43c7aeafed5ff0598244782,
4776fa910633c8b0e76cc775ccad21498a761cef
and 064a5cba30fc5314c37f8c08f43426059e9861e4.

This is an unfortunate situation and not really a proper
solution.  The proper solution would be to setup
~/.wine/drive_c/.windows-serial with a serial number
the matches the real drive serial number.  This was not
enough in some old Wine versions to make TI work,
but apparently it is now.  Unfortunately we have quite
a bit of users who have registered TI with our old
Wine patch, and the new method would be incompatible
with the old one, breaking TI installations.

The current plan is to install another Wine
on the side and support both methods.
After some period of time (2-3 years?) we will remove
this Wine version (with this patch).
Note that old TI installations can also be fixed to move
on to the new scheme, but this requires communication
with TI (the corporation) on each case.