Remove Wine from build.
Building the Stretch version on Buster is not trivial.
Building the 32-bit version on 64-bit system is less trivial.
Yet that is what we need, and only for the legacy TI Nspire CX CAS
case.

Solution: just get our old patched Wine deb from our archive.
We are not going to do that after Buster anyway, or should at
least have a plan to get rid of it.