Allow null as a puavo-conf value in puavo-conf profile files.
This seems strange, but the point is that "ignore this value here"
should be a valid option in /state/etc/puavo/local/puavo_conf.json,
editing through puavo-laptop-setup.  It is not the same as
leaving that particular key out of the file, because then
there is no reasonable way to figure out if a default value
should be used as a policy or if it has simply never been decided
on.