Use the master kerberos server as the default one instead of the (possible) bootserver kerberos (slave).  Fatclients still use the bootserver kerberos only.
This is because there are networks where the DNS will not return
"not found" in time, leading to login failures.  The connection
to the master kerberos should always happen through the vpn,
including the dns call, meaning that it should be reliable as long
as the vpn is working.  But use the possibly existing bootserver
kerberos as a fallback option anyway in case we are on some school
network and the master kerberos server can not be accessed.