Fix jetpipe for Buster.
Brief history of (our) jetpipe: it comes from LTSP project files.
The LTSP (I think) did not handle (and probably still does not
handle) the case where a usb printer is turned on after the thin/fat
client has been booted.  Jetpipe is not started as udev events
occur, only at boot time.

We had our own udev-wrapper for this... we listen to udev events
with usb printers connecting, and then start up jetpipe for those.

This got broken with the systemd version in Buster.  The udev-trigger
would start up jetpipe... and then after about 4-5 seconds, systemd
would kill it, because that is no longer the way udev-triggers should
be written I guess (not for long running processes).  This has
probably been the way for a longer time, and perhaps the jetpipe
had fork() to protect itself from this.  That fork() is longer
sufficient though, systemd looks at process cgroups and kills
the jetpipe process in Buster.

The solution involved writing an udev trigger that tells systemd
it should start a process for the new usb device.  This means
that systemd also knowns when to kill the jetpipe process,
after the device has been detached, making the puavo-jetpipe
wrapper for jetpipe a bit simpler.