puavomenu: New icon locator system
This recursively scans theme-specific directories, so you don't have
to list them all in dirs.json. It also has a sophisticated (yeah,
right) algorithm for sorting the directories for prioritizing
different icon sizes.

I wanted to replace this with the Gtk.IconTheme class, but it did not
produce exactly the results we want. I'd have to keep our own icon
locator as a fallback and if you have to do that, why use Gtk.IconTheme
at all? You'd be doing double-scans and slowing down the startup even
more.

when tested, this produced identical (or better) results for all icon
themes currently enabled in our system.