Add "staff" to ACLs.
I think this is really needed only for the /v3/my_school_users case.
But the current cucumber tests expect that these folks should be
able to do this and that.  We have "admin" for these, I think
"staff" should be mostly equivalent to teachers in terms of
access rights.  But we will see if we change this.