Allow removing of kerberos principals to users who have permission to manage users in an organisation.
Note that this allows removal of principals even across
normal boundaries, for example an admin in some school
can remove a principal for an organisation owner.
These should be caught in the application level.

Removal of service principals is not permitted.

This allows to remove users more cleanly.  Currently
when removing users, we mess up user kerberos passwords so
that those user accounts can not be used.