Make all UEFI-booting devices hybrid MBR/UEFI-bootable.
We still do not have any UEFI-only devices in production
(yes, in late 2020!), so we can still do this without separately
supporting the UEFI-only mode.

I do not know if this will trigger some issues, but we will see.
The main reason for this is that we want "puavo-make-install-disk"
to make removable, bootable usb disks that can be booted both
in legacy BIOS mode and in UEFI.  But then I do not see why we
could not do that with all disks, not just the removable ones.

(Some tricky issues might develop, for example system will have
UEFI-boot, then it is changed to legacy, and the UEFI-grub stops
updating, and then system is switched again to UEFI-boot,
booting with not-up-to-date Grub-version... but do these situations
really happen in the real world and does that matter, that is
another question).