Change "puavo.grub.efi.force_extra_removable" to "true" by default.
This is a tricky question.  But so far our experience has been
that this is needed so widely it needs to be on.  This has
consequences, mark my words, sky is falling, etc.  (Perhaps
nobody should or will ever install dual-boot devices with UEFI?
Or perhaps Microsoft is nice and will not overwrite stuff?
Or perhaps next month I will have a world tour as a rock star?)