In puavo-ers-applet, throw away the (freshly written) session handling code.  I thought I needed it, but the logic for "is exam finished?" or rather "can we treat the exam as finished?" is now simpler and more effective I think, that the benefits of the session saving code is so little that I think it does not deserve to be.  Now that was an interesting roundabout.
I prefer simpler logic here if it is good enough.  I hope this is.

(What this session saving functionality would now provide to us,
feature-wise, would be a bit less of answer saving (in case the exam
time has not changed between puavo-ers-applet process invocations),
and we could also know when to reload exams even if the exam set has
not changed, in case the scheduled end time was over (and I am not
even sure this matters)).