Use meta-files to lookup image/rdiff sizes and checksums.
Use "sha512" as the checksum algorithm for images and rdiffs.
More secure, and on modern machines almost as fast as "cksum".
Not using "sha256", because "sha512" is actually faster, and
that matters to us.

The old "CKSUMS"-file mechanism is no longer used.  We need
to support that for quite some years though, so that all old
machines have an upgrade path to newer versions.