puavomenu: Don't crash if the button label is not a string
I wasn't sure if this should be committed, but it's a nasty situation
that can appear out of nowhere. Better be on the lookout than crash.

The crash happens if you forget to localize button labels and pass
in a dict instead of an str. Yes, a progammer's own fault, but it
has happened in production.