Use separate puavo-conf variables for torrent-based update ratelimiting.
This allows using a higher rate limit for torrent-uploads/-downloads
than for ordinary non-torrent updates.  Set the default value
on this one to 20 megabytes per second (five times the current
default for non-torrent updates).

Also, because with torrents it is possible to use a separate
rate limit for uploads and downloads, make that possible as well.