Pass rate limit information to puavo-torrent-updated.
The point of this is not to use the provided rate limit in
puavo-torrent-updated, but rather, to know if rate limit should
be used.  Normally updates should use rate limits, but it should
be possible to override this on individual cases by admins/users who
want updates to happen very fast.  With this change,
puavo-torrent-updated can refrain from applying rate limits to
torrent downloads/uploads.