Use /var/lib/puavo-desktop/shared for managed printers list in puavo-dnssd-printer-client, so that puavo-dnssd-printer-client could be used by normal users.  Move "puavo-dnssd-printer-client" from /usr/sbin to /usr/bin.
The plan is to enable this for fatclients for getting printers.
We should not need two separate mechanisms for getting printer
information, for fatclients and others.  But this means that
printer restrictions should now be implemented in some other way,
which is not yet done here.  On the other hand, that mechanism
could be used both by fatclients and laptops.