Stop using puavo-template in setup_gdm.
It does not work in the early boot stage in Bullseye.
"/dev/stdin" is missing at that point.  There would be two
options: to fix puavo-template or simply remove it altogether.

I do not think puavo-template proved to be as useful as I thought
it would be.  It has existed for years but has been invoked only
in four places, and after this change, only in two.  Templating
with shell (or some other mechanism) is not that difficult,
so I think it is time for it to go.

(The point of puavo-template was to provide a simple templating
language that would have the puavo-conf variables ready for use).