Fix login screen logo height.
In previous Gnome versions the logo height was fixed to 48 pixels,
and the logo was scaled accordingly.  The current Gnome versions
do not do this scaling, and our logos are way too big.

I guess we could also change the logo image files.  But
fixing the logo height to 48 pixels suited us okay,
so we can continue doing that at least for a while.