In puavo-bootserver-sync-images, delete files even in case sync has failed.
The plan is to do cleanups first before new images are updated.
In some cases the updates do not work because there is not enough
diskspace, and if enough diskspace can not be made by removing
old cruft, admins have to resolve the situation manually.  This
should not be necessary, because images that are needed should not
be cleaned up anyway, as those remain as images to be synchronized.
Besides, the system images (used by the bootserver itself) should
remain in place always.