Fix Gnome Shell from restarting every 90 seconds.
On Bullseye systemd apparently checks for Gnome Shell if it is
running, and restarts it if not.  Our "gnome-shell"-wrapper
messed up its heuristic to find that out, meaning Gnome Shell
was restarting every 90 seconds.

Use "exec" when starting Gnome Shell, and adjust "adm-x" to that.
This fixes the issue.