Do not build our own Grub in Bullseye.
The initial reason to build own Grub was that needed to patch it.
Then, in Buster, we did not need to patch it, but we wanted to
fix it to a particular version, because we know that upgrades
could be unreliable.

Grub updates should now be somewhat more reliable (because shutdowns
during it are inhibited), so perhaps we can simply use the Debian
Grub.

This may need to be revisited.  Nevertheless, it does not make
much sense to build software only to fix it to a particular version.