Revert "Enable mdns lookups in nsswitch.conf."
This reverts commit cfd95e6ba25d5207e12220cabc9df986f294c88f.

Reverting this.   Not because it is wrong (it is not), but
because we do not have any real-world positive/negative
test results regarding this yet, and we want to avoid risks
(anything that suddenly starts working when it previously
has not is also a risk).

This will come back though, but with other changes and
when it is better tested.