puavomenu: Change how string translation works
STRINGS was created so everything can be in one place and to avoid
global variables. But it is a global variable anyway, and accessing
it means you have to haul around the current language everywhere.
Change that to one global "translator" instance.

I spent three hours testing everything and things worked, but I fear
I might still have missed something and the whole thing breaks
in production.