Remove "debian_repository" definition from Puppet-rules.
We have not used this in Buster.  I think it was to include
packages from multiple Debian versions.  But it is not much
better if we can do without including older or newer Debian
versions in apt sources, as we have managed to do in Buster.
One reason why that is bad is that "apt update" takes a much
longer time and all those package lists take more space (that
is not normally a problem, but sometimes might be).  Besides,
mixing packages from several release versions can be problematci.