Use "lpadmin" to set printer destination permissions only if the printer exists.
Otherwise "lpadmin" will create a new printer destination,
which is not managed by cups-browsed or anything else,
and this is not what we want.

This solution has a TOCTOU issue, but there is no
"do-not-create-but-modify-only" feature in lpadmin and apparently
there is no other similar tool to solve the problem,
so this is only a "good enough" solution.