Support hybrid MBR/GPT partitions on puavo-setup-filesystems.
Also add --create-hybrid-mbr option to puavo-make-install-disk.
This should make installation disks that can be booted both
with legacy BIOS and UEFI.  The previous mechanism could do that,
but that only worked on some BIOSes.  Some other BIOSes can not
handle GPT partitions on legacy mode, thus write both the MBR
and GPT partition information to disk.

This should be okay for installation images.  I would not use
this scheme for real installations, and currently it is not possible
to do so.  However the relavant partitioning code was intentionally
added to puavo-setup-filesystems, in case we ever want to create
real USB drive installations that are flexible enough to boot on
all kinds of machines regardless of their booting mode.