Disable Veyon service when "Close a Veyon session"-button has been pushed.
This simply tells clients that master is no longer available,
and they should react accordingly.  Master socket will remain
open, however, in case someone wants to try to connect to it.
The PIN is unset, however, so no actual connections are possible.

This relies on clients to remove their public keys, and correctly
behaving clients should do so.  The master will not remove
its keys, the private/public key may be reused.