Remove our current implementation for getting VirtualBox into the image.
This used the Oracle packages and installed guest additions
as well.  I do not think Oracle quite does a good job on
this (it rebuilds kernel modules on every boot, wasting time
and CPU resources).  I have tried to fight this but the
issue still sometimes exists (Oracle does not realize
that system state might disappear after a reboot).

It appears there are better packages in fasttrack.debian.net,
so we might just as well try those.  These are backported packages
from Debian unstable, and seem like they are done properly
with the dkms system.