In puavo-user-registration, if we have received timeout when communicating with server, disable all other inputs except "previous page" and "submit" on the account page. This state will be kept until another attempt at user creation is done.  The "previous page" is good to keep enabled in case user wants to try another network.
The point of this is that in case of a timeout, user registration
might have been successful but we just do not know it yet.
If user tries again with the same username + password and
primary user for device matches as well, client can receive
a success message on retries and continue.