Clean up lots of old cruft from parts/devscripts.
I am not 100% sure none of this is used somewhere,
but mostly these are not used or usable in our current
infrastructure, so throw these away, they are only
a source of confusion.

The remaining stuff might be moved elsewhere in case
we want to get rid of "puavo-devscripts" altogether.