Make exam sorter also function on demand.
It is unclear if exam sorter should be on its own thread.
It could just be run synchronously after exam sync.
However, in the original design it was planned to make
it possible to have the sorter work without exam sync
(by making a zip file for all exams manually added to
the system).  It no longer works in this way and is
not used in this way, so maybe it should not be run
in its own thread.