Replace case statement with a device selection of 'plughw:CARD=PCH'.
This seems to work on multiple device models and generations.
Had to add a '-p 256' flag to 'speaker-test' to avoid a "Broken pipe" -error.
In addition to enabling the full use of speakertest on EliteBook G1/G2's, the tests are in result more "generalized", at the moment showing no need to workaround some device type specifically.

Tested on EliteBooks (G1 through G5) and some Dells.