Add some insurance into update-configuration.
Now, if puavo-conf-update crashes while handling the new
device.json (coming from Puavo), it will NOT put it into
"/state/etc/puavo/device.json".  Because if it were to be put
into the filesystem, puavo-conf-update would use that on the
next boot and crash, leaving the system into an unusable state.

This provides us some extra safety.  The logic is that
if puavo-conf-update returns, even with some error code,
it has been able to handle the input at least in some reasonable
way.

We could later add some error codes that are fatal in
a similar way as a crash is, to provide safety against some
other possible failure scenarios.