Disable suspend on hosts with "ers"-profile (Examination Room Server).
This is problematic because if host is not on power, battery will
eventually run out and then that is a bigger problem than suspend.
On laptops that have a server-role, we could perhaps treat
the battery as an UPS (uninterruptible power supply) that should
start making some alarm noise if the machine is not connected
to power.  Yet we do not have that yet.

In any case, this is the setting that has actually been in production
for a while and what I hear is that there are less problems with
this than without, so this goes in.