Do not use DNS in fatclient /etc/krb5.conf to look up the kerberos server.
This drops a few DNS lookups on fatclients during login,
which is normally not an issue, except that those DNS requests
go to upstream DNS servers that may be slow to respond,
blocking login.  As we know the kerberos server to use
(the bootserver), simply use that one.

Note that /etc/krb5.conf.masteronly does not contain the
master server on fatclients, only the bootserver.
It is not supposed to work like that (it is used with
external login in some situations), but fatclients do
not know the address of the actual kerberos master,
because the bootserver (through DNS) lies it is the master.