Add a gocryptfs home directory password change hook to our password change operation.
This adds a --pass-form-response-to command line option
to puavo-webwindow, that can be used to pass form parameters
and response information to a script.  "change-homecrypt-password"
then handles the home directory password change operation.

This also requires a change to puavo-web so that webkit
receives a message that password change was successful.