Help cups-browsed a little even on laptops.
Cups-browsed does store the "requesting-user-name-allowed" key
for printer queues, so that it can remember and fill it in later
in case a printer disappears and then reappears.  BUT!  It looks
like in case "AllowUser" gets removed from a printer queue
(giving permissions for all to print), cups-browsed does not realize
the "requesting-user-name-allowed" should be forgotten as well.

With this patch, those stale printer restrictions get cleaned up
after a laptop reboot.  Not perfect, but better.  This should
fix some printer visibility issues, in case some printer restrictions
were previously in place.