Use maximum connect retry of five seconds on our maintenance OpenVPN tunnel.
We have desktop logins going on in the tunnel, and would prefer
to keep the possible connection time short.  This is especially
relevant when puavo-user-registration is used, because there
is a short time between network connection and login, and
this may happen after some time after boot.  We do not want
to raise the OpenVPN restart pause time high because of this
reason.  The resources used are negligible and the extra log
messages a small bother.