When using x11vnc with Veyon, send updates less frequently.
The Veyon use-case does not require as high frame rates,
and this uses quite a bit less network resources and actually
results in better responsiveness when screen updates are frequent.
This may seem counter-intuitive, but when mirroring a screen
to several computers, in case the network can not keep up
with the updates, computers may start lagging behind even
a minute or more.  Veyon should try to do some adaptive
framerate technique (maybe it tries it but I do not see it
working), and maybe better VNC compression algorithms might help
as well.