Do not mess up with X keymap tables in Veyon.
In case veyon-server is restarted while there is a lock or
a demo mode is in fullscreen, Veyon session is broken (as it
should), but veyon-server does not clean up the situation
and leaves users with broken keyboard state (no keys work).
This might happen due to a bug (in case there is a crash),
or intentionally (as in our case if connection is closed
through puavo-veyon-applet).

After this change, the locking behaviour is still mostly
as it is before.  There is a grabKeyboard() call which
perhaps does mostly what is needed.  In case there is
some loophole that allows a user to escape a lock,
I think that is just fine.