Implement pam_puavorolegroups.so functionality in a different way.
The problem with pam_puavorolegroups.so is that while it adds
a role group into a user session, from the outside it looks like
user does not belong to a role group.  This means that dbus
policy rules regarding that group are not effective.

Instead of using the recently written module, check out
user role from a shell script and manage the user groups
with "usermod".