Fix getting kerberos ticket from lock screen in Bullseye.
Behaviour in Bullseye is different from Buster in that
something is renewing the kerberos ticket for user,
but only to the "renew until"-time.  Thus we need to
do a real kerberos login in case lock screen is used
over several days.  The old "check kerberos cache timestamp"
trick does not longer work in Bullseye, because the cache
is occasionally rewritten when ticket is renewed.

The code in this commit checks out the ticket expiration date
and if it is within the next two and a half days,
PAM is directed to do an actual network login.