Update encryption types for kerberos clients.
Include a new one "aes256-cts-hmac-sha384-192", that we
do not use anywhere yet, but should.  The "aes256-cts-hmac-sha1-96"
we have had a long time, it is time to start using it.
The "des3-hmac-sha1" should go away, but we still might have some
servers that do not accept much else so do not remove it yet
(unfortunately).  The "des-cbc-crc" can be removed as current
MIT Kerberos no longer supports it.