Add --convert-to-hdb in puavo-restore-all-databases and apply the old hdb-configurations when converting.
This is only a fallback option in case we get into some trouble
with the mdb format.  It is unlikely we need to ever use this but
I feel better when this option is around.  Once the conversion
to mdb is done this code can be thrown away.

I plan to drop support for hdb in the ldap init scripts, though.