Configure the "lastbind" slapd-module.
This module updates user-specific timestamp each time user
logs into ldap.  With this configuration the module is activated,
but set to update the timestamp only once in an hour.
This should ensure that there are not too many writes to ldap
because of this (each authentication does not trigger a write
operation).

The reason for enabling this module is to track user accounts
which have not been used in a long time.  In Puavo it is not
enough to enable this module because kerberos authentications
to ldap do not leave such a mark, but I think we can do something
about that.