Add configurations for the auditlog-overlay as well.
THe purpose of this is to make it possible to track changes
happening in Puavo database.  More specifically, who changed,
what changed and when it changed.

The log files are organisation specific.  In case the necessary
directories are missing, it appears this module/configuration
does nothing.  But we should test this more.