Fix usb stick formatting from nautilus in bullseye. The problems are apparently caused by a non-numeric xid parameter format used in Bullseye. The upstream has fixed this at https://gitlab.gnome.org/GNOME/gnome-disk-utility/-/merge_requests/40 by completely ignoring the xid parameter.
It should already be in 3.38.2 that Bullseye ships
https://gitlab.gnome.org/GNOME/gnome-disk-utility/-/commit/140baa1bf7b954106ce8d318518d39cb916ae815
However, inspecting
http://deb.debian.org/debian/pool/main/g/gnome-disk-utility/gnome-disk-utility_3.38.2.orig.tar.xz
reveals that although the meson.build and NEWS contain the changed version number, the
/src/disks/gduapplication.c contains old code, e.g. line 175 is
{"xid", 0, 0, G_OPTION_ARG_INT, NULL, N_("Parent window XID for the format dialog"), "ID" }
which is not part of upstream 3.38.2. So, seems like Debian version 3.38.2 somewhy uses partially
sources from 3.38.1.

Let's add a wrapper script that drops xid parameters ourselves.