Enable zram only on fatclients.
Laptops and bootservers have their own swap partitions, unlike
fatclients.  It seems the swap partitions are underutilized
with zram is used.  It is probably better to keep the zram
only in situations where we need it, as fatclients do not usually
not have a disk to put the swap partition on.