In puavo-dump-all-databases, do not bind to ldap to get a list of databases to dump.  Instead, look the list of databases from the cn=config.ldif file.
This is simpler (less code, less dependencies), and has the
benefits that is database format independent (hdb, mdb or whatever
future brings), and more importantly, allows dumping to work
even when slapd is shut down.  This is good, so that database
format changes and such can be safely made with: shutdown slapd,
then dump and restore, startup slapd.