Exclude /puavo-os from image, but add a log history of latest commits into /etc/puavo-image/puavo-os-git.log.
The /puavo-os takes about one percent of space from the
final image and is practically not used for any purpose.
The main benefit of it has been, that it can be used to
check what are latest changes/fixes this particular image
has (and thus what it is missing).  But a simple snippet
of git log history fills that purpose.