Add switch --add-efi-label to puavo-install-grub.
This can be used to add an EFI label in case we know it
might be missing from host.  Previously we had code
to do this, but it was missing an argument -d so that
it worked only with /dev/sda.  As puavo-install-grub knows
about boot devices (and normally sets up the EFI label
as part of grub-install), it makes sense to do this there.