Enable gnome-shell extensions in gdm.
This used to be a one-liner in gnome-shell code plus dconf configuration
for extensions to enable.  Now this is a bit more complicated,
as gnome-shell supports enabling extensions in "gdm"-, "unlock-dialog"-
and "user"-modes (as documented in
https://gjs.guide/extensions/topics/session-modes.html#example-usage).
Extensions need to specify which modes they support.

This is a good idea, however, at the moment none of the extensions
we use in login screen actually support running in login screen
(even though they have already worked there for some years).
Also, for some reason I needed to explicitly enable extensions
(in gnome-shell code) in "gdm"-mode for them to actually work,
even though we already configure those in dconf configuration.
It may be that I am missing something, nevertheless this makes
our login screen work in Bookworm (Gnome 43) much more like
in Bullseye (Gnome 3.38).