Add a bug workaround to DING.
If a user is logged on a device and the only monitor is unplugged, DING will
begin to fill syslog with "DING: Not enough space to add icons" messages.
After /state is full, gvfsd-trash will also start to busyloop, apparently due to
DING polling Trash on desktop with every loop and file handles not being
released, or something similar.

The existence of this bug should be tested with upstream DING and Bookworm GNOME.
Various loops in DING desktopManager.js assume the existence of at least one
display, and although there have been some changes since our version, my first
my first impression based on the code is that this will likely happen there, too.
However, having a logged-in user on a displayless desktop computer is not that
common usecase, and if one VNCs to a such computer, it can be seen that the state
is not too coherent anyhow (gnome-shell wouldn't want to restart in such
conditions). So this is just damage control for a corner case that is probably
not of general interest.