In puavo-reset-laptop-to-factory-defaults, use "shred" instead of "srm".
The reason for this is that srm refuses to overwrite file contents
and remove it in case the file is hard linked to some other path.
There is no option argument or an environment variable to change
this behaviour.  Thus its use in this context is ruled out,
and we switch to shred.