In bootserver installation, install an image before setting up bootserver state.
In case something goes wrong when initializing bootserver state,
the system can still be booted and the initializations can
be done later (by hand).  There are potential downsides to this,
(bootserver installation goes wrong and nobody notices early
enough), but it is usually nice to get a bootable system as early
as possible.