Remove "Pinta" from Bookworm.
This used to be buggy and is missing from Debian Bookworm.
We used to compile this by ourselves.  However, there is a
new version that might be better.  We need to think a bit if
we want to go there or drop this altogether.  Maybe drop as
we have other painting programs available, but we will see.