darkdm: Print some extra information when the device seems to be in UEFI-mode.
Tell the user if it *looks like* the device has our BIOS configs set or not.
This is based on observations on which efivars exist with BIOS factory defaults
and which do not. This method is very crude and probably works only for HP devices,
though others have not even been tested.