Add app-ids to our PWA information.
I would rather not do this, but these should be quite stable
and the old mechanism of looking of from the Chrome
"Preferences"-file no longer works with the latest version of
Google Chrome.  This means that these might occasionally get
broken.  We'll see how often, some kind of hotfix or better
mechanism might be needed.