Start using "dns_canonicalize_hostname = false" in kerberos configurations.
This is needed for SSO Nextcloud access when reverse DNS is
not setup properly.  The old "rdns = false" setting was not
effective enough and it is not effective when this new setting
is in place.

I do not think there should be any regressions due to this,
but we will see.