Add a puavo-conf profile for Nextcloud.
The idea is to be able to simply set puavo.profile.nextcloud --> true
for bootservers that need to be Nextcloud servers, and then
bootservers should install/configure Nextcloud to be ready for use.

The Docker image versions for Nextcloud/Postgres/Collabora are
set here.  This means they should be kept up-to-date.
The versions can be overridden on Puavo, of course, in case
of a need for a rapid patch or some other reason.