When backing up Docker, handle cases where Nextcloud and PostgreSQL are not running.
One case that is not handled currently is one where Nextcloud
is running but PostgresSQL is not.  Nextcloud does not work
in this situation and can not be put into maintenance mode.
As the backup script fails, Nextcloud upgrades are also disabled.
This situation needs manual intervention, I think that is okay
(this is in ShouldNotHappen-territory).