Add /etc/dkms/no-autoinstall file to prevent automatic installation of dkms modules.
We want to install kernels from backports, and also install
some dkms-modules that are not compatible with backport kernels.
Without this file the packaging system tries to compile
modules for all kernels, that does not work.  We specify
ourselves (with puppet rules) which kernel+dkms combinations
we want and work.