Drop a custom build on qt from Bookworm.
Our custom build of qt was never built on Bookworm.  It appears it is also
unnecessary: judging from https://gitlab.gnome.org/GNOME/gimp/-/issues/789
and https://codereview.qt-project.org/c/qt/qtbase/+/306771 the relevant
patch is also in 5.15.8 and fixes the issue ("Successfully created
cherry-pick to tqtc/lts-5.15", "I applied that patch to Qt 5.15 and
tested, can confirm that it fixes the issue").  Also I can not reproduce
the bug after testing, but that may just my bad luck.  I also verified
that this patch is indeed in the current Debian version.