puavomenu: category buttons will receive the initial focus
Compared to giving the inital focus to the search field, this has a
clear benefit on touchscreen devices: the virtual keyboard does not
pop up immediately when the menu is shown. And because pressing any
symbol key already automatically focuses the search field, the "fast
search and spawn" behavior is not lost.