debs/linux/debian: change ABI name to 0.puavo1
I didn't bother to check whether the ABI actually changed, but because
we added a bunch of patches, I wanted to be on the safe side and made
it clear that this is different than 0.deb12 which is built by Debian.

ABI name is relevant for DKMS modules, which in this case means that
they need to be re-built specifically for this kernel.

```
sed -i -r 's/^abiname:.*$/abiname: 0.puavo1/' debian/config/defines
debian/rules debian/control-real
git add debian/*0.puavo1*
git rm debian/*0.deb12*
```