Add an interface to libpuavoconf for opening puavo-conf database directly.
Use this in puavo-conf-daemon and wait until the database open succeeds.
The puavo-conf-daemon should always open the database directly, because
it is the process that implements the dbus-interface, and thus it does
not make sense for it to open the database using the dbus interface.
In case some process has opened the puavo-conf database directly while
puavo-conf-daemon starts, this might happen and then all accesses to
puavo-conf fail that can not use the direct access (only operations by
root-user may succeed).

With these changes this issue is much less serious.  However,
we should avoid running anything accessing puavo-conf as root
during boot to not interfere with puavo-conf-daemon startup.
In case there is a need for such a process, it might be better
for it to have an interface for enforcing the puavo-conf access
through dbus (currently we lack such an interface).