Fix scrolling in puavo-laptop-setup log.
I think it is better to scroll to the bottom after each
new line is output, otherwise progress can not be seen
or user has to scroll down manually.  I think the intention
of the code was this but idle_add(text_view_scroll_to_bottom)
does not get called except maybe at the very end, unless this
this change is applied.