puavo-reset-windows: clear the dirty flag of win primary partitions when using force
Windows primary partition might have dirty flag set (due to ungraceful
shutdown etc.), which is a signal for Windows to run chkdsk on the
next boot. But, when reseting such partition, we don't care if such
filesystem check is scheduled or not, we are going to reset it
anyways.

But for other partitions than the Windows primary partition, we want
to keep the dirty flag, if it's set.

Also, dirty flag is cleared only when --force is used.

So, use the --force Mikko!