When removing puavo-pkgs, purge them completely if they are locally installed.
In case a puavo-conf key for a puavo-pkg package is empty,
and the puavo-pkg is locally installed, unconfigure it and
purge.  This means that for puavo-pkg packages to be installed
locally, the puavo-conf key *must* be set to a value.
In case there is a system puavo-pkg package (in the system image),
switch to that before purging.