Do not allow grub-efi-amd64-signed to be installed.
Apparently this was marked as "essential" package, and removing it may
break the system.  For us, installing it breaks the system.  Ironic.

We need a fix to the bug in
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=959425 before we can
use this.  There is a fix in Ubuntu and Fedora, I think, which we could
apply, but then the package will not be authorized by Microsoft and
those who follow.

Did that sound political?  I did not intend to.