Remove BringOutSubmenuOfPowerOffLogoutButton@pratap.fastmail.fm.
This extension has some issues: tooltips go on top even though
our panel is at bottom.  The tooltip messages are not localized.
Only the buttons that this extension manages have tooltips
but others on the side do not.

However, this extension has a nice idea, so we might bring it
back after these issues are dealt with.  We just do not know yet.
The default Gnome behaviour might be something we prefer, so
we'll test with that for a while.