Dropping the puavo-autopilot releasetest session.
There are some good ideas in here, but there is only one
test which has not been needed in years, and this is not used
for anything, so it is time for this to go.  Can be resurrected
of course if we actually start doing automated QA.  It might also
be that we start doing that with completely different tools.