parts/puavomenu: relaunch on resolution change
Because resolution changes are quite common. But why this is handled
in `launch` command? Because it's the safest and easiest way to ensure
EVERYTHING get's re-initialized correctly with new dimensions in both
dev and prod modes. New runtime, new everything.

Quitting and restarting Gtk main loop does not give good results. And
destroying the existing PuavoMenu and creating a new one when
resolution is changed did not seem a safe option either. And
dynamically changing dims is definitely not a good idea, the whole
reason why we have these semi-static dimensions is that we create
everything once based on the common set and then just live with it.

So, restarting the whole runtime seemed the best option.